// src/front/src/api/fincosts/models/documentEntryListFilter.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PBX Fincosts API (via Gateway)
 * OpenAPI spec version: 1.0
 */

export interface DocumentEntryListFilter {
	pageNumber?: number;
	pageSize?: number;
	readonly skip?: number;
	/** @nullable */
	search?: string | null;
	/** @nullable */
	documentTypeId?: number | null;
	/** @nullable */
	currencyNamePL?: string | null;
	/** @nullable */
	wasPaid?: boolean | null;
}