// src/front/src/api/fincosts/models/documentEntryResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PBX Fincosts API (via Gateway)
 * OpenAPI spec version: 1.0
 */

export interface DocumentEntryResponse {
	/** @nullable */
	gid?: string | null;
	/** @nullable */
	currencyNamePL?: string | null;
	/** @nullable */
	currencyForeignRate?: number | null;
	totalAmount?: number;
	vatRate?: number;
	wasPaid?: boolean;
	/** @nullable */
	kseRefNumber?: string | null;
	/** @nullable */
	kseAccountNumber?: string | null;
	documentTypeId?: number;
	/** @nullable */
	documentTypeNamePL?: string | null;
	/** @nullable */
	issuedBy?: string | null;
	/** @nullable */
	issuedVatNumber?: string | null;
	/** @nullable */
	issuedFor?: string | null;
	/** @nullable */
	issuedForVatNumber?: string | null;
	createdAt?: string;
	/** @nullable */
	modifiedAt?: string | null;
}