// src/front/src/api/fincosts/models/currencyTypeResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PBX Fincosts API (via Gateway)
 * OpenAPI spec version: 1.0
 */

export interface CurrencyTypeResponse {
	id?: number;
	/** @nullable */
	namePL?: string | null;
	/** @nullable */
	nameEN?: string | null;
	/** @nullable */
	category?: string | null;
}